add auth middleware for edit
