+++ components/AuthButton.server.js
@@ -5,19 +5,18 @@ import EditButton from './EditButton.client'
 export default function AuthButton({ children, login, noteId, ...props }) {
   if (login) {
     return (
-      <Link href={`/note/edit/${noteId || ''}`}>
-        <a className="link--unstyled">
-          <EditButton {...props} login={login}>
-            {children}
-            <img
-              src={`https://avatars.githubusercontent.com/${login}?s=40`}
-              alt="User Avatar"
-              title={login}
-              className="avatar"
-            />
-          </EditButton>
-        </a>
-      </Link>
+      // Use hard link
+      <a href={`/note/edit/${noteId || ''}`} className="link--unstyled">
+        <EditButton {...props} login={login}>
+          {children}
+          <img
+            src={`https://avatars.githubusercontent.com/${login}?s=40`}
+            alt="User Avatar"
+            title={login}
+            className="avatar"
+          />
+        </EditButton>
+      </a>
     )
   }
